requirements.txt
numpy<2
pandas
pytorch==2.0.1
pytorch-cuda==11.4
scikit-learn
torchvision
